Agregando cambios de tipo de moneda

+++ src/main/kotlin/com/ucb/edu/abc/mscompany/dto/response/ExchangeRateDto.kt
@@ -3,6 +3,7 @@ package com.ucb.edu.abc.mscompany.dto.response
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.ucb.edu.abc.mscompany.entity.ExchangeMoneyEntity
 import lombok.Data
+import java.math.BigDecimal
 
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -13,7 +14,7 @@ data class ExchangeDateDto(
 
 data class CurrencyDateDto(
         val exchangeRateId: Int,
-        val currency: Double,
+        val currency: BigDecimal,
         val moneyName: String,
         val abbreviationName: String
 )
